fix issue #238 (#247)
* change int to bigint for subject_id

* fix on postgres

+++ R/cohort_ddl.R
@@ -12,8 +12,8 @@ createCohortTables <- function(con, writeSchema, name, computeAttrition) {
                        name = inSchema(writeSchema, name),
                        fields = c(
                          cohort_definition_id = "INT",
-                         # subject_id = "BIGINT",
-                         subject_id = "INT",
+                         subject_id = "BIGINT",
+                         # subject_id = "INT",
                          cohort_start_date = "DATE",
                          cohort_end_date = "DATE"))
